old release working
Wip5 better fullscreen mode runs more games 2d games are playable but slow as before. but it  is a bit smoother. 3d games are not and only really non demanding 2d games such as prism land story is 10 fps and playable turn on frameskip with the keyboard to gain little bit more speed. bios screens now  do  show  copyright warning screen on ridgeracer. i cleaned the  code as much as i could it now compiles  with little to no warnings. the GPU plugin was converted to use SDL_SWSURFACE  re did the store ques part and defined some new modes not used yet.


press page up or down to get to FS< then press home and end till you have a star at like this > *FS< then it will be as fast as it can go for now.


remember its only a wip